requirements.txt
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
h11==0.14.0
idna==3.10
inflection==0.5.1
MouseInfo==0.1.3
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyJWT==2.9.0
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
pytweening==1.2.0
pytz==2025.1
PyYAML==6.0.2
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
